+++ tests/other/org/jgroups/tests/UnicastTestTcpSlow.java
@@ -129,7 +129,7 @@ public class UnicastTestTcpSlow {
         Message msg=new BytesMessage(null, buf.array());
         // msg.writeTo(output);
 
-        ByteArrayDataOutputStream dos=new ByteArrayDataOutputStream((int)(msg.size()));
+        ByteArrayDataOutputStream dos=new ByteArrayDataOutputStream(msg.size());
         byte flags=0;
         dos.writeShort(Version.version); // write the version
         if(msg.getDest() == null)
